refactor(cli): flatten doctor and upgrade command modules
