+++ turbo/apps/cli/src/commands/zero/connector/custom/create.ts
@@ -12,7 +12,7 @@ import {
   connectorActionUrl,
   printCallbackActionUrlExample,
 } from "../action-url";
-import { getPlatformOrigin } from "../../doctor/platform-url";
+import { getPlatformOrigin } from "../../../doctor/platform-url";
 import { createCustomConnectorDefinitionFileSchema } from "./definition";
 
 interface CreateOptions {
