+++ turbo/apps/cli/src/commands/zero/connector/status.ts
@@ -3,7 +3,7 @@ import chalk from "chalk";
 import { listZeroConnectorCatalogStatus } from "../../../lib/api/domains/zero-connectors";
 import { withErrorHandler } from "../../../lib/command/with-error-handler";
 import { resolveAgentContext } from "./agent-context";
-import { getPlatformOrigin } from "../doctor/platform-url";
+import { getPlatformOrigin } from "../../doctor/platform-url";
 import {
   availableConnectorSlugs,
   findConnectorStatusItem,
